Move things and add route changer
