+++ pages/map/deleteNode.php
