+++ pages/map/uploadImage.php
